[DAD-557] feat: 스크랩 카드 추상화
