+++ src/components/molcules/OtherScrapCard.tsx
@@ -25,99 +25,37 @@ interface OtherScrapCardProps {
     }
 }
 
-function OtherScrapCard({ content }: OtherScrapCardProps) {
-    const scrapCardMenu = [{
-        name: '카드 수정하기',
-        onClick: () => {
-            hideTooltip();
-            showScrapEditModal();
-        },
-    }, {
-        name: '카드 삭제하기',
-        onClick: () => {
-            hideTooltip();
-            showScrapDeleteModal();
-        },
-    }];
-
-    const [isTooltipVisible, setIsTooltipVisible] = useState(false);
-    const [isScrapEditModalVisible, setIsScrapEditModalVisible] = useState(false);
-    const [isScrapDeleteModalVisible, setIsScrapDeleteModalVisible] = useState(false);
-    const [isMemoCreateModalVisible, setIsMemoCreateModalVisible] = useState(false);
-
-    function showTooltip() {
-        setIsTooltipVisible(true);
-    }
-
-    function hideTooltip() {
-        setIsTooltipVisible(false);
-    }
-
-    function showScrapEditModal() {
-        setIsScrapEditModalVisible(true);
-    }
-
-    function hideScrapEditModal() {
-        setIsScrapEditModalVisible(false);
-    }
-
-    function showScrapDeleteModal() {
-        setIsScrapDeleteModalVisible(true);
-    }
-
-    function hideScrapDeleteModal() {
-        setIsScrapDeleteModalVisible(false);
-    }
-
-    function showMemoCreateModal() {
-        setIsMemoCreateModalVisible(true);
-    }
-
-    function hideMemoCreateModal() {
-        setIsMemoCreateModalVisible(false);
-    }
+function OtherScrapCard({ content, showMemoCreateModal, showTooltip }: OtherScrapCardProps) {
 
     return (
-        <CardContainer>
-            <CardWrapper
-                style={{ cursor: 'pointer' }}
-                onClick={(e) => {
+        <CardWrapper
+            style={{ cursor: 'pointer' }}
+            onClick={(e) => {
+                e.stopPropagation();
+                window.open(`${content.pageUrl}`);
+            }}
+        >
+            <CardImage src={content.thumbnailUrl} />
+            <CardInfoWrapper>
+                {content.title && <EmpasizedTypography>{content.title}</EmpasizedTypography>}
+                {content.description && <DefaultTypography>{content.description}</DefaultTypography>}
+            </CardInfoWrapper>
+            {content.memoList?.map(memo => {
+                return <Memo memoImageURL={memo.memoImageURL} memoText={memo.memoText} />
+            })}
+            <ButtonContainer>
+                <Button buttonStyle={'gray'} label={'메모 추가하기'} fullWidth isRound onClick={(e) => {
                     e.stopPropagation();
-                    window.open(`${content.pageUrl}`);
-                }}
-            >
-                <CardImage src={content.thumbnailUrl} />
-                <CardInfoWrapper>
-                    {content.title && <EmpasizedTypography>{content.title}</EmpasizedTypography>}
-                    {content.description && <DefaultTypography>{content.description}</DefaultTypography>}
-                </CardInfoWrapper>
-                {content.memoList?.map(memo => {
-                    return <Memo memoImageURL={memo.memoImageURL} memoText={memo.memoText} />
-                })}
-                <ButtonContainer>
-                    <Button buttonStyle={'gray'} label={'메모 추가하기'} fullWidth isRound onClick={(e) => {
-                        e.stopPropagation();
-                        showMemoCreateModal();
-                    }} />
-                    <MoreIconContainer src={MoreIcon} onClick={(e) => {
-                        e.stopPropagation();
-                        showTooltip();
-                    }} />
-                </ButtonContainer>
-            </CardWrapper>
-            {isTooltipVisible && <Tooltip contents={scrapCardMenu} color={theme.color.background_color} />}
-            {isScrapEditModalVisible && <ScrapEditModal hideScrapEditModal={hideScrapEditModal} content={content} />}
-            {isScrapDeleteModalVisible && <ScrapDeleteModal hideScrapDeleteModal={hideScrapDeleteModal} scrapId={content.scrapId} />}
-            {(isScrapEditModalVisible || isScrapDeleteModalVisible || isMemoCreateModalVisible) && <Overlay />}
-            {isMemoCreateModalVisible && <MemoCreateModal hideMemoCreateModal={hideMemoCreateModal} scrapId={content.scrapId} />}
-        </CardContainer>
+                    showMemoCreateModal();
+                }} />
+                <MoreIconContainer src={MoreIcon} onClick={(e) => {
+                    e.stopPropagation();
+                    showTooltip();
+                }} />
+            </ButtonContainer>
+        </CardWrapper>
     );
 }
-
-const CardContainer = styled.div`
-    position: relative;
-    border-radius: 4px;
-`
 
 const CardWrapper = styled.div`
     padding: 15px;
@@ -162,13 +100,4 @@ const MoreIconContainer = styled.img`
     cursor: pointer;
 `
 
-const Overlay = styled.div`
-    width: 100vw;
-    height: 100vh;
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 1;
-    background-color: rgba(0, 0, 0, 0.5);
-`
 export default OtherScrapCard;
