+++ src/components/molcules/VideoScrapCard.tsx
@@ -39,57 +39,7 @@ interface VideoScrapCardProps {
     }
 }
 
-function VideoScrapCard({ content }: VideoScrapCardProps) {
-    const scrapCardMenu = [{
-        name: '카드 수정하기',
-        onClick: () => {
-            hideTooltip();
-            showScrapEditModal();
-        },
-    }, {
-        name: '카드 삭제하기',
-        onClick: () => {
-            hideTooltip();
-            showScrapDeleteModal();
-        },
-    }];
-
-    const [isTooltipVisible, setIsTooltipVisible] = useState(false);
-    const [isScrapEditModalVisible, setIsScrapEditModalVisible] = useState(false);
-    const [isScrapDeleteModalVisible, setIsScrapDeleteModalVisible] = useState(false);
-    const [isMemoCreateModalVisible, setIsMemoCreateModalVisible] = useState(false);
-
-    function showTooltip() {
-        setIsTooltipVisible(true);
-    }
-
-    function hideTooltip() {
-        setIsTooltipVisible(false);
-    }
-
-    function showScrapEditModal() {
-        setIsScrapEditModalVisible(true);
-    }
-
-    function hideScrapEditModal() {
-        setIsScrapEditModalVisible(false);
-    }
-
-    function showScrapDeleteModal() {
-        setIsScrapDeleteModalVisible(true);
-    }
-
-    function hideScrapDeleteModal() {
-        setIsScrapDeleteModalVisible(false);
-    }
-
-    function showMemoCreateModal() {
-        setIsMemoCreateModalVisible(true);
-    }
-
-    function hideMemoCreateModal() {
-        setIsMemoCreateModalVisible(false);
-    }
+function VideoScrapCard({ content, showMemoCreateModal, showTooltip }: VideoScrapCardProps) {
 
     const videoMenus = [{
         title: '게시일',
@@ -103,108 +53,87 @@ function VideoScrapCard({ content }: VideoScrapCardProps) {
     },]
 
     return (
-        <CardContainer>
-            <CardWrapper
-                style={{ cursor: 'pointer' }}
-                onClick={(e) => {
-                    e.stopPropagation();
-                    window.open(`${content.pageUrl}`);
-                }}>
-                {content.siteName && <Chip>{content.siteName}</Chip>}
-                {content.title && <EmpasizedTypography>{content.title}</EmpasizedTypography>}
-                <VideoPlayer src={content.embedUrl} />
-                {
-                    content.channelName &&
-                    <RowContainer style={{ alignItems: 'center', gap: '5px' }}>
-                        {content.channelImageUrl ? <ProfileImage size={24} source={content.channelImageUrl} /> : <ProfileImage size={24} source={defaultImage} />}
-                        <DefaultTypography>{content.channelName}</DefaultTypography>
-                    </RowContainer>
-                }
-                <RowContainer style={{ justifyContent: 'space-between' }}>
-                    {videoMenus.map(menu => {
-                        return (
-                            <>
-                                {menu.content &&
-                                    <ColumnContainer style={{ alignItems: 'center', flex: '1' }}>
-                                        <EmpasizedTypography>{menu.content}</EmpasizedTypography>
-                                        <DefaultTypography>{menu.title}</DefaultTypography>
-                                    </ColumnContainer>
-                                }
-                            </>
-                        )
-                    })}
+        <CardWrapper
+            style={{ cursor: 'pointer' }}
+            onClick={(e) => {
+                e.stopPropagation();
+                window.open(`${content.pageUrl}`);
+            }}>
+            {content.siteName && <Chip>{content.siteName}</Chip>}
+            {content.title && <EmpasizedTypography>{content.title}</EmpasizedTypography>}
+            <VideoPlayer src={content.embedUrl} />
+            {
+                content.channelName &&
+                <RowContainer style={{ alignItems: 'center', gap: '5px' }}>
+                    {content.channelImageUrl ? <ProfileImage size={24} source={content.channelImageUrl} /> : <ProfileImage size={24} source={defaultImage} />}
+                    <DefaultTypography>{content.channelName}</DefaultTypography>
                 </RowContainer>
-                {content.description && <DefaultTypography>{content.description}</DefaultTypography>}
-                {content.memoList?.map(memo => {
-                    return <Memo memoImageURL={memo.memoImageUrl} memoText={memo.memoText} />
+            }
+            <RowContainer style={{ justifyContent: 'space-between' }}>
+                {videoMenus.map(menu => {
+                    return (
+                        <>
+                            {menu.content &&
+                                <ColumnContainer style={{ alignItems: 'center', flex: '1' }}>
+                                    <EmpasizedTypography>{menu.content}</EmpasizedTypography>
+                                    <DefaultTypography>{menu.title}</DefaultTypography>
+                                </ColumnContainer>
+                            }
+                        </>
+                    )
                 })}
-                <ButtonContainer>
-                    <Button buttonStyle={'gray'} label={'메모 추가하기'} fullWidth isRound onClick={(e) => {
-                        e.stopPropagation();
-                        showMemoCreateModal();
-                    }} />
-                    <MoreIconContainer src={MoreIcon} onClick={(e) => {
-                        e.stopPropagation();
-                        showTooltip();
-                    }} />
-                </ButtonContainer>
-            </CardWrapper>
-            {isTooltipVisible && <Tooltip contents={scrapCardMenu} color={theme.color.background_color} />}
-            {isScrapEditModalVisible && <ScrapEditModal hideScrapEditModal={hideScrapEditModal} content={content} />}
-            {isScrapDeleteModalVisible && <ScrapDeleteModal hideScrapDeleteModal={hideScrapDeleteModal} scrapId={content.scrapId} />}
-            {(isScrapEditModalVisible || isScrapDeleteModalVisible || isMemoCreateModalVisible) && <Overlay />}
-            {isMemoCreateModalVisible && <MemoCreateModal hideMemoCreateModal={hideMemoCreateModal} scrapId={content.scrapId} />}
-        </CardContainer>
+            </RowContainer>
+            {content.description && <DefaultTypography>{content.description}</DefaultTypography>}
+            {content.memoList?.map(memo => {
+                return <Memo memoImageURL={memo.memoImageUrl} memoText={memo.memoText} />
+            })}
+            <ButtonContainer>
+                <Button buttonStyle={'gray'} label={'메모 추가하기'} fullWidth isRound onClick={(e) => {
+                    e.stopPropagation();
+                    showMemoCreateModal();
+                }} />
+                <MoreIconContainer src={MoreIcon} onClick={(e) => {
+                    e.stopPropagation();
+                    showTooltip();
+                }} />
+            </ButtonContainer>
+        </CardWrapper>
     );
 }
 
-const CardContainer = styled.div`
-    position: relative;
-    border-radius: 4px;
-`
-
 const CardWrapper = styled.div`
-    padding: 15px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    background-color: white;
-    border-radius: 4px;
-`
+            padding: 15px;
+            display: flex;
+            flex-direction: column;
+            gap: 10px;
+            background-color: white;
+            border-radius: 4px;
+            `
 
 const VideoPlayer = styled.iframe`
-    border-radius: 4px;
-`
+            border-radius: 4px;
+            `
 
 const EmpasizedTypography = styled.span`
-    font-size: 20px;
-    font-weight: bold;
-    color: ${theme.color.text_gray_color};
-`
+            font-size: 20px;
+            font-weight: bold;
+            color: ${theme.color.text_gray_color};
+            `
 
 const DefaultTypography = styled.span`
-    font-size: 14px;
-    color: ${theme.color.text_gray_color};
-`
+            font-size: 14px;
+            color: ${theme.color.text_gray_color};
+            `
 
 const ButtonContainer = styled.div`
-    display: flex;
-    align-items: center;
-`
+            display: flex;
+            align-items: center;
+            `
 
 const MoreIconContainer = styled.img`
-    width: 24px;
-    height: 24px;
-    cursor: pointer;
-`
+            width: 24px;
+            height: 24px;
+            cursor: pointer;
+            `
 
-const Overlay = styled.div`
-    width: 100vw;
-    height: 100vh;
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 1;
-    background-color: rgba(0, 0, 0, 0.5);
-`
 export default VideoScrapCard;
